+++ engine/src/dpx_tt_glyf.rs
@@ -349,7 +349,7 @@ pub unsafe extern "C" fn tt_add_glyph(
         & 1i32 << 7i32 - new_gid as i32 % 8i32
         != 0
     {
-        warn!("Slot {} already used.", new_gid as i32);
+        warn!("Slot {} already used.", new_gid);
     } else {
         if (*g).num_glyphs as i32 + 1i32 >= 65534i32 {
             _tt_abort(b"Too many glyphs.\x00" as *const u8 as *const i8);
